refactor: dedup element/tool leftovers — #294 + #215 C2/C3/C9
Five small behavior-preserving extractions, closing out the non-combat
dedup backlog:

- wall/box _setup_textures → GameElement._setup_mirror_textures
  (uv_offset param carries box's rare-variant palette row) [#294]
- telepipe/warp_point _build_mesh → GameElement._build_warp_cylinder
  (tint + node name params; #215-C3 note was stale — they extend
  GameElement, not WarpBase, so the helper lives there)
- bear/needle trap _setup_materials → GameElement._setup_split_materials
  (feature-texture match splits prong/spike vs base material) [#215-C2]
- inventory_screen._get_item_category deleted — line-for-line copy of
  the canonical Inventory.get_item_category [#215-C9]
- point_in_triangle → scripts/tools/quest_geom.py, imported by
  generate_field_quest.py + simulate_field_quest.py [#294]

Existing coverage exercises all five (element tests, inventory
categorization, field-quest sim); no new behavior to characterize.
Dup baseline 131 → 126 (0 new); suite 1528 passed / 0 failed.
