+++ export_presets.cfg
@@ -183,7 +183,7 @@ architectures/arm64-v8a=true
 architectures/x86=false
 architectures/x86_64=false
 version/code=1
-version/name="0.33.3"
+version/name="0.33.4"
 package/unique_name="com.pszgodot.game"
 package/name="PSZ Godot"
 package/signed=true
