+++ scripts/3d/elements/box.gd
@@ -5,8 +5,6 @@ class_name Box
 
 signal destroyed_box
 
-const MIRROR_SHADER = preload("res://scripts/3d/shaders/mirror_repeat.gdshader")
-
 ## Whether this box is a rare variant (uses o0c_recont instead of o01_cont)
 @export var is_rare: bool = false
 
@@ -44,7 +42,7 @@ func _ready() -> void:
 	super._ready()
 	collision_body = _build_static_collision("BoxCollision")
 	_setup_hurtbox()
-	_setup_textures()
+	_setup_mirror_textures(Vector2(0, 1) if is_rare else Vector2.ZERO)
 	_setup_reticle()
 
 
@@ -61,25 +59,6 @@ func _setup_hurtbox() -> void:
 	shape.position.y = hurtbox_size.y / 2
 	hurtbox.add_child(shape)
 	add_child(hurtbox)
-
-
-func _setup_textures() -> void:
-	if not model:
-		return
-	apply_to_all_materials(func(mat: Material, mesh: MeshInstance3D, surface: int):
-		if mat is StandardMaterial3D:
-			var std_mat := mat as StandardMaterial3D
-			if std_mat.albedo_texture:
-				var smat := ShaderMaterial.new()
-				smat.shader = MIRROR_SHADER
-				smat.set_shader_parameter("albedo_texture", std_mat.albedo_texture)
-				smat.set_shader_parameter("uv_scale", Vector2(2, 2))
-				smat.set_shader_parameter("mirror_x", true)
-				smat.set_shader_parameter("mirror_y", true)
-				if is_rare:
-					smat.set_shader_parameter("uv_offset", Vector2(0, 1))
-				mesh.set_surface_override_material(surface, smat)
-	)
 
 
 func _apply_state() -> void:
