+++ scripts/3d/elements/wall.gd
@@ -5,8 +5,6 @@ class_name Wall
 
 signal destroyed_wall
 
-const MIRROR_SHADER = preload("res://scripts/3d/shaders/mirror_repeat.gdshader")
-
 ## Whether this wall can be destroyed by player attacks
 @export var is_destructible: bool = true
 
@@ -23,24 +21,7 @@ func _init() -> void:
 func _ready() -> void:
 	super._ready()
 	collision_body = _build_static_collision("WallCollision")
-	_setup_textures()
-
-
-func _setup_textures() -> void:
-	if not model:
-		return
-	apply_to_all_materials(func(mat: Material, mesh: MeshInstance3D, surface: int):
-		if mat is StandardMaterial3D:
-			var std_mat := mat as StandardMaterial3D
-			if std_mat.albedo_texture:
-				var smat := ShaderMaterial.new()
-				smat.shader = MIRROR_SHADER
-				smat.set_shader_parameter("albedo_texture", std_mat.albedo_texture)
-				smat.set_shader_parameter("uv_scale", Vector2(2, 2))
-				smat.set_shader_parameter("mirror_x", true)
-				smat.set_shader_parameter("mirror_y", true)
-				mesh.set_surface_override_material(surface, smat)
-	)
+	_setup_mirror_textures()
 
 
 func _apply_state() -> void:
