+++ scripts/3d/elements/warp_point.gd
@@ -19,28 +19,9 @@ func _init() -> void:
 
 
 func _ready() -> void:
-	_build_mesh()
+	_mesh = _build_warp_cylinder(Color(0.67, 0.4, 1.0), "WarpPointMesh")
 	_setup_collision()
 	_apply_state()
-
-
-func _build_mesh() -> void:
-	_mesh = MeshInstance3D.new()
-	var cyl := CylinderMesh.new()
-	cyl.top_radius = 0.8
-	cyl.bottom_radius = 0.8
-	cyl.height = 3.0
-	_mesh.mesh = cyl
-	var mat := StandardMaterial3D.new()
-	mat.albedo_color = Color(0.67, 0.4, 1.0, 0.6)
-	mat.transparency = BaseMaterial3D.TRANSPARENCY_ALPHA
-	mat.emission_enabled = true
-	mat.emission = Color(0.67, 0.4, 1.0)
-	mat.emission_energy_multiplier = 2.0
-	_mesh.material_override = mat
-	_mesh.name = "WarpPointMesh"
-	_mesh.position.y = 1.5
-	add_child(_mesh)
 
 
 func _on_interact(_player: Node3D) -> void:
