+++ scripts/tools/simulate_field_quest.py
@@ -23,6 +23,7 @@ import struct
 import sys
 from collections import deque
 from typing import Optional
+from quest_geom import point_in_triangle
 
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 PROJECT_ROOT = os.path.join(SCRIPT_DIR, '..', '..')
@@ -158,15 +159,6 @@ def filter_floor_triangles(
 # Geometry helpers
 # ---------------------------------------------------------------------------
 
-def point_in_triangle(px: float, pz: float, t: tuple) -> bool:
-    """Test if point (px, pz) is inside triangle t = ((x1,z1), (x2,z2), (x3,z3))."""
-    (x1, z1), (x2, z2), (x3, z3) = t
-    d1 = (px - x2) * (z1 - z2) - (x1 - x2) * (pz - z2)
-    d2 = (px - x3) * (z2 - z3) - (x2 - x3) * (pz - z3)
-    d3 = (px - x1) * (z3 - z1) - (x3 - x1) * (pz - z1)
-    has_neg = (d1 < 0) or (d2 < 0) or (d3 < 0)
-    has_pos = (d1 > 0) or (d2 > 0) or (d3 > 0)
-    return not (has_neg and has_pos)
 
 
 def _point_to_triangle_dist_sq(px: float, pz: float, t: tuple) -> float:
